move powerSet to common (#5740)
move powerSet to common: we shall need it in assertions
