+++ kotest-common/src/jvmTest/kotlin/io/kotest/common/PowerSetIndexesTest.kt
@@ -1,10 +1,9 @@
-package com.sksamuel.kotest.property.exhaustive
+package io.kotest.common
 
 import io.kotest.core.annotation.EnabledIf
 import io.kotest.core.annotation.LinuxOnlyGithubCondition
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.collections.shouldContainExactlyInAnyOrder
-import io.kotest.property.exhaustive.powerSetIndexes
 
 @EnabledIf(LinuxOnlyGithubCondition::class)
 class PowerSetIndexesTest: StringSpec() {
@@ -15,14 +14,14 @@ class PowerSetIndexesTest: StringSpec() {
             listOf(0),
          )
       }
-       "should work for two elements" {
-          powerSetIndexes(2).toList() shouldContainExactlyInAnyOrder listOf(
-             listOf(),
-             listOf(0),
-             listOf(1),
-             listOf(0, 1),
-          )
-       }
+      "should work for two elements" {
+         powerSetIndexes(2).toList() shouldContainExactlyInAnyOrder listOf(
+            listOf(),
+            listOf(0),
+            listOf(1),
+            listOf(0, 1),
+         )
+      }
       "should work for three elements" {
          powerSetIndexes(3).toList() shouldContainExactlyInAnyOrder listOf(
             listOf(),
