+++ kotest-property/src/commonMain/kotlin/io/kotest/property/exhaustive/collections.kt
@@ -1,5 +1,6 @@
 package io.kotest.property.exhaustive
 
+import io.kotest.common.powerSetIndexes
 import io.kotest.property.Exhaustive
 
 fun <A> Exhaustive.Companion.collection(collection: Collection<A>): Exhaustive<A> {
@@ -67,20 +68,4 @@ fun <A> Exhaustive.Companion.powerSet(list: List<A>): Exhaustive<List<A>> {
       .exhaustive()
 }
 
-internal fun powerSetIndexes(size: Int): Sequence<List<Int>> = sequence {
-   require(size > 0) { "Size should be positive, was: $size"}
-   val elementsIncluded = MutableList(size) { true }
-   val allIndexes = (0 until size).toList()
-   yield(allIndexes)
-   while(elementsIncluded.any { it }) {
-      for (index in 0 until size) {
-         if (elementsIncluded[index]) {
-            elementsIncluded[index] = false
-            yield(allIndexes.filterIndexed { i, _ -> elementsIncluded[i] })
-            break
-         } else {
-            elementsIncluded[index] = true
-         }
-      }
-   }
-}
+
